move hashing2 functions to their relevant files [NFC]
